ontology_simulator/requirements.txt
fastapi>=0.135.0
uvicorn[standard]>=0.41.0
motor>=3.7.0
pymongo>=4.16.0
python-dotenv>=1.2.0
nats-py>=2.6.0
